Remove `t` slacks from `ocp_nlp_out` (#1148)
The slacks `t` that were used to slack the inequality constraint values
are actually not needed in an NLP solver and are thus removed in this
PR.
Instead of using `t`, we now directly evaluate inequality residual using
the function evaluations.
This results in consistent inequality residuals and will help future
work on globalization.
Related interface functionality such as getters and setters for `t` and
`initialize_t_slacks` have been removed.